+++ src/Entity/SolutionVideoEvent.php
@@ -4,22 +4,14 @@ declare(strict_types=1);
 
 namespace App\Entity;
 
-use App\Entity\Trait\WithUlidCreatedAt;
+use App\Entity\Trait\WithUlid;
 use App\Repository\SolutionVideoEventRepository;
 use Doctrine\ORM\Mapping as ORM;
-use Symfony\Bridge\Doctrine\IdGenerator\UlidGenerator;
-use Symfony\Component\Uid\Ulid;
 
 #[ORM\Entity(repositoryClass: SolutionVideoEventRepository::class)]
 class SolutionVideoEvent
 {
-    use WithUlidCreatedAt;
-
-    #[ORM\Id]
-    #[ORM\CustomIdGenerator(class: UlidGenerator::class)]
-    #[ORM\GeneratedValue(strategy: 'CUSTOM')]
-    #[ORM\Column(type: 'ulid', unique: true)]
-    private ?Ulid $id = null;
+    use WithUlid;
 
     #[ORM\ManyToOne(inversedBy: 'solutionVideoEvents')]
     #[ORM\JoinColumn(nullable: false)]
@@ -28,18 +20,6 @@ class SolutionVideoEvent
     #[ORM\ManyToOne(inversedBy: 'solutionVideoEvents')]
     #[ORM\JoinColumn(nullable: false)]
     private ?Question $question = null;
-
-    public function getId(): ?Ulid
-    {
-        return $this->id;
-    }
-
-    public function setId(Ulid $id): static
-    {
-        $this->id = $id;
-
-        return $this;
-    }
 
     public function getOpener(): ?User
     {
