--- src/Entity/Trait/WithUlidCreatedAt.php
@@ -1,18 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Entity\Trait;
-
-trait WithUlidCreatedAt
-{
-    /**
-     * Get the created at date and time of the solution event.
-     *
-     * @return ?\DateTimeImmutable The parsed DateTime object
-     */
-    public function getCreatedAt(): ?\DateTimeImmutable
-    {
-        return $this->id?->getDateTime();
-    }
-}
